Update saga to fetch products
Sadly, it looks like this method won't work, because dumb shopify
doesn't have any way to fetch metafields, and I need those to
calculate which products to show.

Instead, I should maintain a local map between productId and lat/lng,
and fetch the product(s) that are closest to the location specified.