Refactor cart reducer
The cart reducer now holds an array of cart items, which contain the
minimum data needed to recreate the cart:

* product ID
* variant ID
* quantity

A selector joins this data to the product itself, so that the cart
contents can be rendered.

This commit also incldues some work hydrating persisted carts.
Once the cart is read, product data is fetched from Shopify, so that
we can actually render the items in the persisted cart.